feat: secure cron endpoints and add scheduled workflows
